#4 refactored App component so that the renderApp function uses a switch statement && updated updateEvent function so that it does not use a next ternary expression
